# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  /** @minLength 6 */
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type PropertyType = (typeof PropertyType)[keyof typeof PropertyType];

export const PropertyType = {
  land: "land",
  chalet: "chalet",
  apartment: "apartment",
  villa: "villa",
} as const;

export type PropertyStatus =
  (typeof PropertyStatus)[keyof typeof PropertyStatus];

export const PropertyStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export interface Property {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  price: number;
  location: string;
  locationAr: string;
  city: string;
  cityAr: string;
  type: PropertyType;
  status: PropertyStatus;
  area: number;
  images: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreatePropertyRequestType =
  (typeof CreatePropertyRequestType)[keyof typeof CreatePropertyRequestType];

export const CreatePropertyRequestType = {
  land: "land",
  chalet: "chalet",
  apartment: "apartment",
  villa: "villa",
} as const;

export type CreatePropertyRequestStatus =
  (typeof CreatePropertyRequestStatus)[keyof typeof CreatePropertyRequestStatus];

export const CreatePropertyRequestStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export interface CreatePropertyRequest {
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  price: number;
  location: string;
  locationAr: string;
  city: string;
  cityAr: string;
  type: CreatePropertyRequestType;
  status: CreatePropertyRequestStatus;
  area: number;
  images: string[];
}

export type UpdatePropertyRequestType =
  (typeof UpdatePropertyRequestType)[keyof typeof UpdatePropertyRequestType];

export const UpdatePropertyRequestType = {
  land: "land",
  chalet: "chalet",
  apartment: "apartment",
  villa: "villa",
} as const;

export type UpdatePropertyRequestStatus =
  (typeof UpdatePropertyRequestStatus)[keyof typeof UpdatePropertyRequestStatus];

export const UpdatePropertyRequestStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export interface UpdatePropertyRequest {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  location?: string;
  locationAr?: string;
  city?: string;
  cityAr?: string;
  type?: UpdatePropertyRequestType;
  status?: UpdatePropertyRequestStatus;
  area?: number;
  images?: string[];
}

export interface PropertiesListResponse {
  properties: Property[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Booking {
  id: number;
  name: string;
  phone: string;
  message: string;
  propertyId: number;
  property?: Property;
  createdAt: string;
}

export interface CreateBookingRequest {
  /** @minLength 2 */
  name: string;
  /** @minLength 7 */
  phone: string;
  message: string;
  propertyId: number;
}

export interface BookingsListResponse {
  bookings: Booking[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface Stats {
  totalProperties: number;
  availableProperties: number;
  reservedProperties: number;
  soldProperties: number;
  totalBookings: number;
  recentBookings: number;
}

export type GetPropertiesParams = {
  search?: string;
  type?: GetPropertiesType;
  status?: GetPropertiesStatus;
  minPrice?: number;
  maxPrice?: number;
  page?: number;
  limit?: number;
};

export type GetPropertiesType =
  (typeof GetPropertiesType)[keyof typeof GetPropertiesType];

export const GetPropertiesType = {
  land: "land",
  chalet: "chalet",
  apartment: "apartment",
  villa: "villa",
} as const;

export type GetPropertiesStatus =
  (typeof GetPropertiesStatus)[keyof typeof GetPropertiesStatus];

export const GetPropertiesStatus = {
  available: "available",
  sold: "sold",
  reserved: "reserved",
} as const;

export type GetBookingsParams = {
  page?: number;
  limit?: number;
};