# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Azzam Real Estate API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: properties
    description: Property operations
  - name: bookings
    description: Booking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /properties:
    get:
      operationId: getProperties
      tags: [properties]
      summary: Get all properties with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [land, chalet, apartment, villa]
        - name: status
          in: query
          schema:
            type: string
            enum: [available, sold, reserved]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertiesListResponse"

    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyRequest"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a single property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyRequest"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: Get all bookings (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingsListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/stats:
    get:
      operationId: getStats
      tags: [bookings]
      summary: Get dashboard statistics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{id}:
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
      required:
        - id
        - email
        - name
        - role

    Property:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        location:
          type: string
        locationAr:
          type: string
        city:
          type: string
        cityAr:
          type: string
        type:
          type: string
          enum: [land, chalet, apartment, villa]
        status:
          type: string
          enum: [available, sold, reserved]
        area:
          type: number
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - titleAr
        - description
        - descriptionAr
        - price
        - location
        - locationAr
        - city
        - cityAr
        - type
        - status
        - area
        - images
        - createdAt
        - updatedAt

    CreatePropertyRequest:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        location:
          type: string
        locationAr:
          type: string
        city:
          type: string
        cityAr:
          type: string
        type:
          type: string
          enum: [land, chalet, apartment, villa]
        status:
          type: string
          enum: [available, sold, reserved]
        area:
          type: number
        images:
          type: array
          items:
            type: string
      required:
        - title
        - titleAr
        - description
        - descriptionAr
        - price
        - location
        - locationAr
        - city
        - cityAr
        - type
        - status
        - area
        - images

    UpdatePropertyRequest:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        location:
          type: string
        locationAr:
          type: string
        city:
          type: string
        cityAr:
          type: string
        type:
          type: string
          enum: [land, chalet, apartment, villa]
        status:
          type: string
          enum: [available, sold, reserved]
        area:
          type: number
        images:
          type: array
          items:
            type: string

    PropertiesListResponse:
      type: object
      properties:
        properties:
          type: array
          items:
            $ref: "#/components/schemas/Property"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - properties
        - total
        - page
        - limit
        - totalPages

    Booking:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        message:
          type: string
        propertyId:
          type: integer
        property:
          $ref: "#/components/schemas/Property"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - message
        - propertyId
        - createdAt

    CreateBookingRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 7
        message:
          type: string
        propertyId:
          type: integer
      required:
        - name
        - phone
        - message
        - propertyId

    BookingsListResponse:
      type: object
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - bookings
        - total
        - page
        - limit
        - totalPages

    Stats:
      type: object
      properties:
        totalProperties:
          type: integer
        availableProperties:
          type: integer
        reservedProperties:
          type: integer
        soldProperties:
          type: integer
        totalBookings:
          type: integer
        recentBookings:
          type: integer
      required:
        - totalProperties
        - availableProperties
        - reservedProperties
        - soldProperties
        - totalBookings
        - recentBookings