# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { Property } from "./property";

export interface Booking {
  id: number;
  name: string;
  phone: string;
  message: string;
  propertyId: number;
  property?: Property;
  createdAt: Date;
}