# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  /** @minLength 2 */
  name: string;
  /** @minLength 7 */
  phone: string;
  message: string;
  propertyId: number;
}