# lib/api-zod/src/generated/types/propertiesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { Property } from "./property";

export interface PropertiesListResponse {
  properties: Property[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}