# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from "./propertyStatus";
import type { PropertyType } from "./propertyType";

export interface Property {
  id: number;
  title: string;
  titleAr: string;
  description: string;
  descriptionAr: string;
  price: number;
  location: string;
  locationAr: string;
  city: string;
  cityAr: string;
  type: PropertyType;
  status: PropertyStatus;
  area: number;
  images: string[];
  createdAt: Date;
  updatedAt: Date;
}