# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalProperties: number;
  availableProperties: number;
  reservedProperties: number;
  soldProperties: number;
  totalBookings: number;
  recentBookings: number;
}