# lib/api-zod/src/generated/types/updatePropertyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Azzam Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePropertyRequestStatus } from "./updatePropertyRequestStatus";
import type { UpdatePropertyRequestType } from "./updatePropertyRequestType";

export interface UpdatePropertyRequest {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  location?: string;
  locationAr?: string;
  city?: string;
  cityAr?: string;
  type?: UpdatePropertyRequestType;
  status?: UpdatePropertyRequestStatus;
  area?: number;
  images?: string[];
}